+++ CHANGES.md
@@ -1,3 +1,8 @@
+### 20240425
+ * CM script automation: fixed dumping versions (variations didn't have _ prefix)
+
+---
+
 Since March 2023, all updates to CM automations are submitted via PRs.
 You can follow our PRs at
 * https://github.com/ctuning/mlcommons-ck/commits/master